bug-014: quest completed-block key ushort->uint — daily reset clears high-id quests (t_8b47a3bf)
CharacterQuests keyed completed blocks by ushort (block = questId/64); for
quest ids >= 4,194,304 the key wrapped (8000004 -> 125000 -> 59464). Storage
and reads wrapped consistently, but ResetQuests recomputed questId =
completeBlockId*64 + blockIndex = 3,805,700 -> GetTemplate null -> the daily
completed bit was never cleared -> AddQuest refused with QuestDailyLimit
forever. Live carrier: 8000004 (할로윈 축제 준비).

- CompletedQuest.Id ushort -> uint; CompletedQuests dict key uint
- HasQuestCompleted/SetCompletedQuestFlag/IsQuestComplete use uint block math
- Load reads id via GetUInt32 (column is INT UNSIGNED — no SQL change)
- SCCompletedQuestsPacket casts to ushort at the wire boundary (1.2 client
  wire format unchanged); SCQuestContextResetPacket already uint
- Regression pin: HighIdDailyQuest_Completed_ResetDailyQuests_ClearsFlag_And_Reaccepts
  (QuestDailyResetRigTests) — completes 8000004, reset clears the flag,
  engine AddQuest re-accepts
- Wi11bProbeTests (TEMP probe) removed; census T16 8000004 flips to PASS,
  4573/4573 quests runnable, 0 FAIL (runnability.md regenerated)
- Full gate: 1494 passed / 0 failed / 1 env-gated skip ([LiveRigOnly])

+++ AAEmu.Game/Models/Game/Char/CharacterQuests.cs
@@ -22,7 +22,10 @@ public class CharacterQuests(Character owner)
 
     private Character Owner { get; set; } = owner;
     public Dictionary<uint, Quest> ActiveQuests { get; } = [];
-    private Dictionary<ushort, CompletedQuest> CompletedQuests { get; } = [];
+    // BUG-014: completed blocks are keyed by uint (block id = questId / 64) — a
+    // ushort key wrapped for quest ids >= 4,194,304 and ResetQuests recomputed a
+    // wrapped quest id that never resolved to a template (bit never cleared).
+    private Dictionary<uint, CompletedQuest> CompletedQuests { get; } = [];
 
     public bool HasQuest(uint questId)
     {
@@ -31,7 +34,7 @@ public class CharacterQuests(Character owner)
 
     public bool HasQuestCompleted(uint questId)
     {
-        var questBlockId = (ushort)(questId / 64);
+        var questBlockId = questId / 64;
         var questBlockIndex = (int)(questId % 64);
         return CompletedQuests.TryGetValue(questBlockId, out var questBlock) && questBlock.Body.Get(questBlockIndex);
     }
@@ -385,7 +388,7 @@ public class CharacterQuests(Character owner)
     public CompletedQuest SetCompletedQuestFlag(uint questId, bool isCompleted)
     {
         // Calculate block and index
-        var completedQuestBlockId = (ushort)(questId / 64);
+        var completedQuestBlockId = questId / 64;
         var completedQuestBlockIndex = (ushort)(questId % 64);
         // Grab or create block
         if (!CompletedQuests.TryGetValue(completedQuestBlockId, out var completedBlock))
@@ -405,7 +408,7 @@ public class CharacterQuests(Character owner)
     /// <returns></returns>
     public bool IsQuestComplete(uint questId)
     {
-        var completeId = (ushort)(questId / 64);
+        var completeId = questId / 64;
         if (!CompletedQuests.TryGetValue(completeId, out var completedQuest))
             return false;
         return completedQuest.Body[(int)(questId % 64)];
@@ -509,7 +512,7 @@ public class CharacterQuests(Character owner)
                 {
                     var quest = new CompletedQuest
                     {
-                        Id = reader.GetUInt16("id"),
+                        Id = reader.GetUInt32("id"),
                         Body = new BitArray((byte[])reader.GetValue("data"))
                     };
                     CompletedQuests.Add(quest.Id, quest);
